Time: 0 ms (100%), Space: 7.9 MB (88.69%) - LeetHub

+++ 0062-unique-paths/0062-unique-paths.cpp
@@ -32,7 +32,7 @@ public:
         // return prev[n-1];
 
         // combinatorics ncr
-        double res = 1;
+        long long res = 1;
         int N = m+n-2;
         int r = m-1;
         for(int i=1; i<=r; i++){
